shared: Remove WriteTempFile
