+++ shared/osarch/tempfile_test.go
@@ -1,19 +1,23 @@
 package osarch
 
 import (
+	"io/ioutil"
 	"os"
 
 	"github.com/stretchr/testify/suite"
-
-	"github.com/lxc/lxd/shared"
 )
 
 // WriteTempFile writes content to a temporary file.
 func WriteTempFile(s suite.Suite, dir string, prefix string, content string) (string, func()) {
-	filename, err := shared.WriteTempFile(dir, prefix, content)
+	f, err := ioutil.TempFile(dir, prefix)
 	if err != nil {
-		s.T().Errorf("failed to create temporary file: %v", err)
+		s.T().Errorf("Failed to create temporary file: %v", err)
 	}
+	defer f.Close()
 
-	return filename, func() { os.Remove(filename) }
+	_, err = f.WriteString(content)
+	if err != nil {
+		s.T().Errorf("Failed to write string to temp file: %v", err)
+	}
+	return f.Name(), func() { os.Remove(f.Name()) }
 }
